api/email/requirements.txt
cowpy==1.1.5
python-dotenv==0.20.0
sendgrid==6.9.7
